Add files via upload
Using data of a telecom industry, wanted to find the probability of customer churn in the company based on the attributes available. Hence built a logistic regression model on the data to find out probability of customer churn in the industry and found out what are the factors driving customer churn the most. The model showed 85% true probability of customer churn and also showed factors influencing customer churn in the industry